Extract mobile gallery modal component
